style file structure update
